// src/store/user/user.types.ts
export enum USER_ACTION_TYPES {
    SET_CURRENT_USER = 'SET_CURRENT_USER',
    CHECK_USER_SESSION = 'user/CHECK_USER_SESSION',
    GOOGLE_SIGN_IN_START = 'user/GOOGLE_SIGN_IN_START',
    EMAIL_SIGN_IN_START = 'user/EMAIL_SIGN_IN_START',
    SIGN_IN_SUCCESS = 'user/SIGN_IN_SUCCESS',
    SIGN_IN_FAILED = 'user/SIGN_IN_FAILED',
    SIGN_OUT = 'user/SIGN_OUT',
    SIGN_UP = 'user/SIGN_UP',
    SIGN_UP_FAILED = 'user/SIGN_UP_FAILED'
}